HACK: Quick hack to throw away any mem=,nvmem=,vmalloc= args
in cmdline, to be replaced with an Atrix-appropriate CONFIG_CMDLINE,
if CONFIG_CMDLINE_PREPEND_ATRIX=y (which should contain the proper
mem=,nvmem=,vmalloc=).  This way, even though the unlockable international
bootloader passes the wrong mem=, we can hack it in .config, and still not
need boot.img unique for each tegrapart, since we now inherit the rest of
cmdline from bootloader.